% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plunder.R
\name{plunder}
\alias{plunder}
\title{Plunder Points From A Plot}
\usage{
plunder(path, labels)
}
\arguments{
\item{path}{Character. The local path to a plot..}

\item{labels}{Logical. Get prompted to add labels to each point you extract?}
}
\value{
A data.frame with two columns (x and y) and a row for each point the
user clicked. Three columns (x, y, labels) if labels is TRUE.
}
\description{
Extract interactively the x and y coordinates from a plot provided as a PNG.
}
\examples{
\dontrun{plunder("example.png")}
}
