% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plunderplot-package.R
\docType{package}
\name{plunderplot-package}
\alias{plunderplot}
\alias{plunderplot-package}
\title{plunderplot: Plunder Points From Images Of Charts}
\description{
Interactively extract x and y point values from a PNG or JPEG image of a chart. Powered by the locator function.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.github.com/matt-dray/plotplunder}
  \item Report bugs at \url{https://www.github.com/matt-dray/plotplunder/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
